authorized CRUD

+++ app/Http/Controllers/CharacterController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Character;
+use http\Env\Response;
 use Illuminate\Http\Request;
 
 class CharacterController extends Controller
@@ -19,29 +20,54 @@ class CharacterController extends Controller
 
     //[Admin] View list of all characters in table
     public function overview(){
+
+        if(auth()->guest() || auth()->user()->username != 'hewwohana'){
+            abort(\Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+        }
+
         $characters = Character::all();
         return view('/admin/overview',['characters' => $characters]);
     }
 
     //[admin] edit form
     public function edit($id){
+
+        if(auth()->guest() || auth()->user()->username != 'hewwohana'){
+            abort(\Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+        }
+
         $character = Character::find($id);
         return view('/admin/edit', ['character' => $character]);
     }
 
     //[admin] add a new character
     public function create(){
+
+        if(auth()->guest() || auth()->user()->username != 'hewwohana'){
+            abort(\Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+        }
+
         return view('admin/addCharacter');
     }
 
     //[admin] details character
     public function read($id){
+
+        if(auth()->guest() || auth()->user()->username != 'hewwohana'){
+            abort(\Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+        }
+
         $character = Character::find($id);
         return view('/admin/read', ['character' => $character]);
     }
 
     //[admin] save new character
     public function store(Request $request){
+
+        if(auth()->guest() || auth()->user()->username != 'hewwohana'){
+            abort(\Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+        }
+
         $character = new Character;
         $character->charName = $request->input('charName');
         $character->charVision = $request->input('charVision');
@@ -55,6 +81,11 @@ class CharacterController extends Controller
 
     //[admin] saves new data of character
     public function update(Request $request, $id){
+
+        if(auth()->guest() || auth()->user()->username != 'hewwohana'){
+            abort(\Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+        }
+
         $character = Character::find($id);
         $character->charName = $request->input('charName');
         $character->charVision = $request->input('charVision');
@@ -66,6 +97,11 @@ class CharacterController extends Controller
 
     //[admin] delete character
     public function destroy($id){
+
+        if(auth()->guest() || auth()->user()->username != 'hewwohana'){
+            abort(\Symfony\Component\HttpFoundation\Response::HTTP_FORBIDDEN);
+        }
+
         $character = Character::find($id);
         $character->delete();
         return redirect()->back()->with('status','Character Deleted Successfully');
